requirements.txt
streamlit
ultralytics
opencv-python
Pillow
pyyaml
